+++ src/gpu_kernels/AtomisticMagSANS_StructRotKernel.h
@@ -50,7 +50,7 @@ void Atomistic_MagSANS_Kernel_StructRot(MagnetizationData MagData,\
 	unsigned long int N_avg = *MagData.N_avg;
 	unsigned long int W = *MagData.TotalAtomNumber;
 
-	//float v = 1.0/((float)  (*MagData.K)) * pow(1.0/((float) (*MagData.N)), 2); // pre factor
+	//float v = 1.0/((float)  (*MagData.K)) * powf(1.0/((float) (*MagData.N)), 2); // pre factor
 	float v =  1.0/((float) W) * 1.0/((float) N_avg);
 
 	int i = blockIdx.x * blockDim.x + threadIdx.x;
@@ -200,9 +200,9 @@ void Atomistic_MagSANS_Kernel_StructRot(MagnetizationData MagData,\
 		SANSData.S_NucMag_2D[i] = 0.0;
 
 		// polarized magnetic SANS cross section projected in the (qz, qy)-plane
-		SANSData.S_Mag_2D_polarized[i] = v * pow(Px, 2) * (Qx_real * Qx_real + Qx_imag * Qx_imag) \
-									   + v * pow(Py, 2) * (Qy_real * Qy_real + Qy_imag * Qy_imag) \
-									   + v * pow(Pz, 2) * (Qz_real * Qz_real + Qz_imag * Qz_imag) \
+		SANSData.S_Mag_2D_polarized[i] = v * powf(Px, 2) * (Qx_real * Qx_real + Qx_imag * Qx_imag) \
+									   + v * powf(Py, 2) * (Qy_real * Qy_real + Qy_imag * Qy_imag) \
+									   + v * powf(Pz, 2) * (Qz_real * Qz_real + Qz_imag * Qz_imag) \
 									   + v * 2.0 * Px * Py * (Qx_real * Qy_real + Qx_imag * Qy_imag) \
 									   + v * 2.0 * Px * Pz * (Qx_real * Qz_real + Qx_imag * Qz_imag) \
 									   + v * 2.0 * Py * Pz * (Qy_real * Qz_real + Qy_imag * Qz_imag);
